Enhance AuthorServiceImp with detailed logging and best practices adaptation
- Introduce LoggerFacade usage across all methods to provide detailed logging for operations such as saving, updating, fetching, and deleting authors. This includes logging successful operations as well as potential errors, significantly improving the service's maintainability and debuggability.

- Adapt existing methods to follow best practices, including the use of @Transactional for database operations, ensuring consistent and reliable transaction management.

- Refine exception handling by throwing specific exceptions like AuthorNotFoundException and AuthorAlreadyExistsException, which enhances error feedback and control flow.

- Utilize ModelMapper for clean and maintainable object mapping between DTOs and entities, reducing boilerplate code and potential for mapping errors.

- Implement getAllPageable and findAllByName methods with logging and exception handling, providing robust functionalities for paginated author retrieval and search by name or lastname, respectively.

These improvements aim to enhance the code quality, readability, and error handling of the AuthorServiceImp, aligning with modern Java best practices and contributing to the overall stability and performance of the application.

Implement comprehensive testing suite for controllers, services, and repository layers

- Developed unit tests for all endpoints in the AuthorController and BookController, ensuring robustness and reliability of the REST API. These tests simulate HTTP requests and assert responses to validate the correctness of request handling, status codes, and output formats.

- Added unit tests for service layers (AuthorServiceImp and BookServiceImp), focusing on the business logic validation, exception handling, and interaction with repository layers. Mocking techniques are utilized to isolate service layer tests from external dependencies, allowing for precise behavior verification.

- Created integration tests for the AuthorRepository and BookRepository to verify the interaction with the database. These tests ensure that JPA queries and custom methods behave as expected when dealing with real database operations, using an in-memory database for a lightweight, isolated testing environment.

- Employed ModelMapper in service layer tests to validate object mapping between DTOs and entities, ensuring data integrity and consistency across layers.

- Enhanced test coverage and reliability by including scenarios such as successful operations, expected failures (e.g., NotFoundException, AlreadyExistsException), and edge cases, significantly reducing the risk of regressions.

- Integrated logging into test executions to provide visibility into the test flow and facilitate debugging of failed tests.

This comprehensive testing suite strengthens the application's overall quality, enabling early detection of issues, facilitating maintenance, and supporting continuous integration and delivery practices.

+++ src/main/java/com/library/libraries/service/imp/BookServiceImp.java
@@ -2,17 +2,21 @@ package com.library.libraries.service.imp;
 
 import com.library.libraries.dto.BookDto;
 import com.library.libraries.dto.BookOneDto;
+import com.library.libraries.dto.BookUpdateDto;
 import com.library.libraries.exception.AuthorNotFoundException;
 import com.library.libraries.exception.BookAlreadyExistsException;
 import com.library.libraries.exception.BookDeletionException;
 import com.library.libraries.exception.BookNotFoundException;
+import com.library.libraries.logging.LoggerFacade;
 import com.library.libraries.model.Author;
 import com.library.libraries.model.Book;
+import com.library.libraries.model.Status;
 import com.library.libraries.repository.AuthorRepository;
 import com.library.libraries.repository.BookRepository;
 import com.library.libraries.repository.UserRepository;
 import com.library.libraries.service.BookService;
 import com.library.libraries.service.util.TPage;
+import com.library.libraries.specification.BookSpecifications;
 import javassist.NotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
@@ -21,6 +25,8 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.validation.Valid;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -33,6 +39,8 @@ public class BookServiceImp implements BookService {
     private final AuthorRepository authorRepository;
     private final BookRepository bookRepository;
 
+    private static final LoggerFacade logger = new LoggerFacade(BookServiceImp.class);
+
     public BookServiceImp(ModelMapper modelMapper, UserRepository userRepository, AuthorRepository authorRepository,
                           BookRepository bookRepository) {
         super();
@@ -42,32 +50,72 @@ public class BookServiceImp implements BookService {
         this.bookRepository = bookRepository;
     }
 
+    @Transactional
     public BookOneDto save(BookOneDto bookOneDto) {
 
+        logger.info("Attempting to save a new book with title: \"{}\"", bookOneDto.getTitle());
+
         if (bookRepository.existsByTitle(bookOneDto.getTitle().trim())) {
+            logger.error("Book already exists with title: \"{}\"", bookOneDto.getTitle());
             throw new BookAlreadyExistsException();
         }
 
         Author author = authorRepository.findById(bookOneDto.getAuthorId())
-                .orElseThrow(AuthorNotFoundException::new);
+                .orElseThrow(() -> {
+                    logger.error("Author not found for ID: {}", bookOneDto.getAuthorId());
+                    return new AuthorNotFoundException();
+                });
 
         Book book = modelMapper.map(bookOneDto, Book.class);
         book.setAuthor(author);
         Book savedBook = bookRepository.save(book);
 
+        logger.info("Book successfully saved with ID: {}", savedBook.getId());
+
         return modelMapper.map(savedBook, BookOneDto.class);
     }
 
-    public List<BookDto> getAll() throws NotFoundException {
+    public List<BookDto> getAll() {
+        logger.info("Retrieving all books");
+
         List<Book> books = bookRepository.findAll(Sort.by(Sort.Direction.ASC, "id"));
+
         if (books.isEmpty()) {
-            throw new NotFoundException("No books available.");
+            logger.info("No books found");
+            return Collections.emptyList();
         }
+
+        logger.info("Found {} books", books.size());
 
         return books.stream()
                 .map(book -> modelMapper.map(book, BookDto.class))
                 .collect(Collectors.toList());
     }
+
+    @Transactional
+    public BookUpdateDto update(Long id, @Valid BookUpdateDto bookUpdateDto) throws NotFoundException {
+        logger.info("Attempting to update book with ID: {}", id);
+
+        Book book = bookRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Book does not exist id: " + id));
+
+        if (book.getStatus() == Status.RENTED) {
+            logger.error("Cannot update book with ID: {} because it is currently RENTED", id);
+            throw new IllegalStateException("Cannot update book as it is currently RENTED.");
+        }
+
+        Author author = authorRepository.findById(bookUpdateDto.getAuthorId())
+                .orElseThrow(() -> new NotFoundException("Author does not exist id: " + bookUpdateDto.getAuthorId()));
+
+        modelMapper.map(bookUpdateDto, book);
+        book.setAuthor(author);
+
+        book = bookRepository.save(book);
+        logger.info("Book with ID: {} successfully updated", book.getId());
+
+        return modelMapper.map(book, BookUpdateDto.class);
+    }
+
 
     public TPage<BookDto> getAllPageable(Pageable pageable) throws NotFoundException {
         Page<Book> page = bookRepository.findAll(pageable);
@@ -85,40 +133,83 @@ public class BookServiceImp implements BookService {
     }
 
     public BookOneDto getOne(Long id) throws NotFoundException {
+        logger.info("Retrieving book with ID: {}", id);
 
-        return bookRepository.findById(id)
+        BookOneDto bookOneDto = bookRepository.findById(id)
                 .map(book -> modelMapper.map(book, BookOneDto.class))
-                .map(bookOneDto -> {
-                    bookOneDto.setAuthorId(bookOneDto.getAuthor().getId());
-                    return bookOneDto;
-                })
-                .orElseThrow(() -> new NotFoundException("Book with ID: " + id + " does not exist."));
+                .orElseThrow(() -> {
+                    logger.error("Book with ID: {} does not exist", id);
+                    return new NotFoundException("Book with ID: " + id + " does not exist.");
+                });
+
+        logger.info("Successfully retrieved book with ID: {}", id);
+        return bookOneDto;
     }
 
     @Transactional
     public void delete(Long id) throws BookNotFoundException, BookDeletionException {
+        Book book = bookRepository.findById(id)
+                .orElseThrow(() -> {
+                    String errorMessage = "Book with ID: " + id + " does not exist.";
+                    logger.error(errorMessage);
+                    return new BookNotFoundException(errorMessage);
+                });
+
+        if (book.getStatus() == Status.RENTED) {
+            String errorMessage = "Cannot delete book with ID: " + id + " as it is currently RENTED.";
+            logger.error(errorMessage);
+            throw new BookDeletionException(errorMessage);
+        }
+
         try {
-            Book book = bookRepository.findById(id).orElseThrow(() -> new BookNotFoundException("Book with ID: " + id + " does not exist."));
             bookRepository.delete(book);
-        } catch (BookNotFoundException e) {
-
-            throw e;
+            logger.info("Successfully deleted book with ID: {}", id);
         } catch (Exception e) {
-
-            throw new BookDeletionException("Error deleting book with id: " + id, e);
+            String errorMessage = "Error deleting book with ID: " + id;
+            logger.error(errorMessage, e);
+            throw new BookDeletionException(errorMessage, e);
         }
     }
 
+    public List<BookDto> searchBooksByTitle(String title) {
 
-    public List<BookDto> searchBooksByTitle(String title) throws NotFoundException {
-        List<Book> books = bookRepository.SearchBooksByTitle(title.trim());
+        logger.info("Searching books with title: \"{}\"", title);
+        List<Book> books = bookRepository.searchBooksByTitle(title.trim());
+
         if (books.isEmpty()) {
-            throw new NotFoundException("No books found with the given title: " + title);
+            logger.info("No books found with the given title: \"{}\"", title);
+            return Collections.emptyList();
         }
+
+        logger.info("{} books found with the given title: \"{}\"", books.size(), title);
 
         return books.stream()
                 .map(book -> modelMapper.map(book, BookDto.class))
                 .collect(Collectors.toList());
     }
 
+    public List<BookDto> searchBooks(BookDto bookDto) {
+
+        try {
+            logger.info("Initiating search for books with criteria");
+
+            List<Book> books = bookRepository.findAll(BookSpecifications.byAttributes(bookDto));
+
+            if (books.isEmpty()) {
+                logger.info("No books found matching the criteria");
+                return Collections.emptyList();
+            }
+
+            logger.info("{} books found matching the criteria", books.size());
+
+            return books.stream()
+                    .map(book -> modelMapper.map(book, BookDto.class))
+                    .collect(Collectors.toList());
+
+        } catch (Exception e) {
+
+            logger.error("Error occurred during book search", e);
+            throw new RuntimeException("Search operation failed", e);
+        }
+    }
 }
